refactored class connection_manager and added logging
